% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_star.R
\name{d.star}
\alias{d.star}
\title{D-star}
\usage{
d.star(C, rho)
}
\arguments{
\item{C}{Covariance matrix}

\item{rho}{Percentage of variation to be explained}
}
\value{
Number d of factors needed to explain rho percent of the variation
}
\description{
This function determines the number d of factors needed to explain percentage rho of the variation induced by covariance matrix C
}
