% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Exponential.cov}
\alias{Exponential.cov}
\title{Exponential.cov}
\usage{
Exponential.cov(Years = 1)
}
\arguments{
\item{Years}{Dimension of the covariance}

\item{Delta}{Resolution in the maturity direction}
}
\value{
Discretized (in resolution Delta) covariance kernel in L^2(0,(Years/Delta)^2)
}
\description{
Calculates discretized product kernel of two exponentials \eqn{k(x,y)\propto e^{-(x+y)}} (normalized to have trace =1 ), where discretization is by local averaging over the grid-cell values
}
