% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functional_Data_Truncation.R
\name{Functional.data.truncation}
\alias{Functional.data.truncation}
\title{Functional Data Truncation}
\usage{
Functional.data.truncation(d, C, data, Delta, sd)
}
\arguments{
\item{d}{Number of inverted eigenvalues}

\item{C}{Statistical covariance matrix}

\item{data}{Data matrix (rows = time)}

\item{Delta}{Increment size}

\item{sd}{Number of standard deviations truncated}
}
\value{
List containing the truncation function values and truncation locations
}
\description{
This function calculates the truncation function for each datum and then determines truncation locations
}
