% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Gaussian.cov}
\alias{Gaussian.cov}
\title{Gaussian.cov}
\usage{
Gaussian.cov(alpha, Years)
}
\arguments{
\item{alpha}{parameter of the gaussian covariance}

\item{Years}{Dimension of the covariance}

\item{Delta}{Resolution in the maturity direction}
}
\value{
Discretized (with resolution Delta) covariance kernel in L^2(0,(days/Delta)^2)
}
\description{
Calculates a discretized Gaussian kernel \eqn{k(x,y)=e^{-alpha (x-y)^2}}, where discretization is by local averaging over the grid-cell values
}
