% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral_Exponential_Kernel.R
\name{integral.Exponential.kernel}
\alias{integral.Exponential.kernel}
\title{Calculator of integral with Gaussian kernel}
\usage{
integral.Exponential.kernel(i, j, Delta = 1/M)
}
\arguments{
\item{i}{Index for first interval}

\item{j}{Index for second interval}

\item{Delta}{Length of each interval}
}
\value{
Integral value
}
\description{
This function calculates the integral of a Gaussian kernel function over a specified interval
}
