% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integral_Gaussian_Kernel.R
\name{integral.Gaussian.kernel}
\alias{integral.Gaussian.kernel}
\title{Calculator of integral with Gaussian kernel}
\usage{
integral.Gaussian.kernel(i, j, alpha = 0.1, Delta = 1/M)
}
\arguments{
\item{i}{Index for first interval}

\item{j}{Index for second interval}

\item{alpha}{Parameter for width of Gaussian kernel}

\item{Delta}{Length of each interval}
}
\value{
Integral value
}
\description{
This function calculates the integral of a Gaussian kernel function over a specified interval
}
