% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2.HS.norm.R
\name{L2.HS.norm}
\alias{L2.HS.norm}
\title{L2.HS.norm}
\usage{
L2.HS.norm(M, from = 0, to = 1)
}
\arguments{
\item{M}{Input matrix to compute the Hilbert-Schmidt norm}

\item{from}{Start of the range}

\item{to}{End of the range}
}
\value{
Scaled L2 Hilbert-Schmidt norm of M
}
\description{
This function computes the L2 Hilbert-Schmidt norm of a matrix M over a specified range
}
