% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2.norm.R
\name{L2.norm}
\alias{L2.norm}
\title{L2.norm}
\usage{
L2.norm(v, from = 0, to = 1)
}
\arguments{
\item{v}{Input vector to compute L2 norm}

\item{from}{Start of the range}

\item{to}{End of the range}
}
\value{
Scaled L2 norm of v
}
\description{
This function computes the L2 norm of a vector v over a specified range
}
