% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2.HS.norm.R
\name{L2_HS_norm}
\alias{L2_HS_norm}
\title{L2 Hilbert–Schmidt Norm}
\usage{
L2_HS_norm(M, from = 0, to = 1)
}
\arguments{
\item{M}{Numeric matrix representing the discretized kernel.}

\item{from}{Start of the interval (default is 0).}

\item{to}{End of the interval (default is 1).}
}
\value{
Numeric value of the approximated \eqn{L^2([from,to]^2)} norm.
}
\description{
Computes the \eqn{L^2([from,to]^2)}-norm of a discretized kernel matrix.
Assumes that the kernel is defined over a square grid of equally spaced points.
}
\examples{
kernel <- outer(seq(0, 1, length.out = 100), seq(0, 1, length.out = 100),
                function(s, t) exp(-abs(s - t)))
L2_HS_norm(kernel)
}
