% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2.norm.R
\name{L2_norm}
\alias{L2_norm}
\title{L2 Norm of a Discretized Function}
\usage{
L2_norm(v, from = 0, to = 1)
}
\arguments{
\item{v}{Numeric vector representing values of a discretized function.}

\item{from}{Start of the interval (default is 0).}

\item{to}{End of the interval (default is 1).}
}
\value{
Numeric value of the approximated \eqn{L^2([from, to])} norm.
}
\description{
Computes the \eqn{L^2([from, to])}-norm of a discretized function (vector).
Assumes the function is evaluated at equally spaced grid points over the interval \code{[from, to]}.
}
\examples{
x <- seq(0, 1, length.out = 100)
f <- sin(pi * x)
L2_norm(f)
}
