% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Simulator}
\alias{Simulator}
\title{Simulator}
\usage{
Simulator(n = 100, lambda = 3, rho = 1, q, k, f0 = numeric(201))
}
\arguments{
\item{n}{Number of simulated days}

\item{lambda}{jump intensity}

\item{rho}{jump magnitude}

\item{q}{cross-sectional covariance kernel of the continuous driver (if 1 year of data should be simulated with m maturity days, the covariance needs to be of dimension m+n, due to simulating the shift at the boundary)}

\item{k}{cross-sectional covariance kernel of the jumps (if 1 year of data should be simulated with m maturity days, the covariance needs to be of dimension m+n, due to simulating the shift at the boundary)}

\item{f0}{initial forward curve}
}
\description{
Simulates from a forward curve model \eqn{df_t= \partial_x f_t + dW_t^Q+dJ_t}, where \eqn{Q} is an integral kernel operator with kernel q and \eqn{J} is a compound Poisson process with intensity lambda and centered gaussian jump distribution with covariance kernel k
}
\examples{
q<-Gaussian.cov(alpha=10,Delta=0.01, Years=2)
}
