% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncated_Covariation_Estimator.R
\name{Truncated.Covariation.estimator}
\alias{Truncated.Covariation.estimator}
\title{Truncated covariation estimator}
\usage{
Truncated.Covariation.estimator(
  adjusted.increments,
  Threshhold.manually = FALSE,
  quantile.adjustment = TRUE,
  truncation = "sparse",
  Prelim.truncation.quantile = 0.75,
  l = 3
)
}
\arguments{
\item{adjusted.increments}{xyz}

\item{Threshhold.manually}{xyz}

\item{quantile.adjustment}{xyz}

\item{truncation}{xyz}

\item{Prelim.truncation.quantile}{xyz}

\item{l}{xyz}
}
\value{
xyz
}
\description{
xyz
}
