% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncated_Covariation.R
\name{truncated_covariation}
\alias{truncated_covariation}
\title{Truncated Covariation Estimator}
\usage{
truncated_covariation(x, tq = 0.75, l = 3, sparse = TRUE, sumplot = FALSE)
}
\arguments{
\item{x}{Matrix of bond prices where x[i,j] is the bond price at the i-th time and j-th maturity (in days).}

\item{tq}{Quantile for truncation in preliminary covariation estimate used for jump truncation.}

\item{l}{Truncation level for functional data truncation (e.g. l=2,3,4,5).}

\item{sparse}{if TRUE, apply B-spline smoothing for high-dimensional maturity data.}

\item{sumplot}{If TRUE, produces summary plots of the results.}
}
\value{
A list with:
  \item{IV}{Truncated covariation matrix}
  \item{locs}{Locations of Detected jumps in the data}
  \item{C.Prel}{Preliminary covariation estimator}
  \item{adj.increments}{Adjusted log-increments}
  \item{expl.var}{Explained variance of eigenvalues}
  \item{gn}{Truncation noise curve}
}
\description{
Computes the (jump-truncated) realized covariation of difference returns \eqn{d_i^n(j) :=  \log(P_{i+1,j}^n)- \log(P_{i,j+1}^n)- \log(P_{i+1,j-1}^n)+ \log(P_{i,j}^n)} from a matrix of daily bond price data \eqn{P_{i,j)}}
}
\references{
Schroers, D. (2024+). Dynamically Consistent Analysis of Realized Covariations in Term Structure Models.
}
