% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variation.R
\name{Variation}
\alias{Variation}
\title{Calculator of the semigroup-adjusted realized covariation (SARCV)}
\usage{
Variation(Incr)
}
\arguments{
\item{Incr}{Increments in form of a numeric matrix}
}
\value{
SARCV as a symmetric matrix
}
\description{
This function takes the increments and calculates the SARCV
}
