% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.cov.R
\name{Exponential.cov}
\alias{Exponential.cov}
\title{Exponential.cov}
\usage{
Exponential.cov(M, n)
}
\arguments{
\item{M}{Number of simulated maturity grid points}

\item{n}{Number of Days}
}
\description{
Calculates a discretized product kernel of two exponentials, where discretization is by local averaging over the grid-cell values
}
