% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian.cov.R
\name{Gaussian.cov}
\alias{Gaussian.cov}
\title{Gaussian.cov}
\usage{
Gaussian.cov(alpha, M, n)
}
\arguments{
\item{alpha}{Dimensionality Parameter}

\item{M}{Number of simulated maturity grid points}

\item{n}{Number of Days}
}
\value{
Number d of factors needed to explain rho percent of the variation
}
\description{
Calculates a discretized Gaussian kernel, where discretization is by local averaging over the grid-cell values
}
