% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\name{Simulator}
\alias{Simulator}
\title{Simulator}
\usage{
Simulator(
  n = 100,
  M = 100,
  lambda = 5,
  rho = 50,
  q,
  k,
  f0 = numeric(n),
  jump.summary.plot = TRUE
)
}
\arguments{
\item{n}{Number of Days}

\item{M}{Number of simulated maturity grid points}
}
\description{
Simulates from a forward curve model
}
